requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cachetools==4.2.4; python_version ~= '3.5'
certifi==2022.6.15; python_version >= '3.6'
chardet==3.0.4
django==1.11.23
ecdsa==0.17.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc]==2.8.2; python_version >= '3.6'
google-auth==2.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-secret-manager==2.11.1
google-crc32c==1.3.0
googleapis-common-protos[grpc]==1.56.3; python_version >= '3.6'
grpc-google-iam-v1==0.12.4; python_version >= '3.6'
grpcio-status==1.47.0
grpcio==1.47.0
idna==2.7
jose==1.0.0
proto-plus==1.20.6; python_version >= '3.6'
protobuf==3.19.4; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-jose==2.0.2
pytz==2022.1
requests==2.20.0
rsa==4.8; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.24.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
